+++ resources/views/components/button/button.blade.php
@@ -2,19 +2,20 @@
     'text',
     'withIcon' => false,
     'iconClass' => '',
-    'iconOnly' => false
+    'iconOnly' => false,
+    'noOpinion' => false
     ])
 
-    <button type="button" {{ $attributes->merge(['class' => 'btn']) }}>
+<button type="button" {{ $noOpinion ? $attributes : $attributes->merge(['class' => 'btn']) }}>
 
-        @if($withIcon)
-            <x-dynamic-component :component="'gt-icon-' .$withIcon" class="icon {{ $iconClass }}" />
-        @endif
+    @if($withIcon)
+        <x-dynamic-component :component="'gt-icon-' .$withIcon" class="icon {{ $iconClass }}" />
+    @endif
 
-        @unless($iconOnly)
-            @isset($text)
-                <span>{{ $text }}</span>
-            @endisset
-        @endunless
+    @unless($iconOnly)
+        @isset($text)
+            <span>{{ $text }}</span>
+        @endisset
+    @endunless
 
-    </button>
+</button>