+++ resources/views/components/table/th.blade.php
@@ -12,20 +12,20 @@
 
         @else
 
-            <button class="flex px-0 bdr-0 fw7 cursor-pointer">
+            <button class="flex px-0 bdr-0 fw7 va-b">
 
                 <span>{{ $slot }}</span>
 
                 @if($direction === 'asc')
-                    <svg class="icon" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                    <svg class="wh-1 txt-red" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7"></path>
                     </svg>
                 @elseif($direction === 'desc')
-                    <svg class="icon" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                    <svg class="wh-1 txt-red" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M5 15l7-7 7 7"></path>
                     </svg>
                 @else
-                    <svg class="icon txt-muted" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
+                    <svg class="wh-1 txt-muted" viewBox="0 0 24 24" xmlns="http://www.w3.org/2000/svg">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M19 9l-7 7-7-7">
                         </path>
                     </svg>